refactor: set graphql as known third party
