+++ src/mkdocstrings_handlers/graphql/_internal/errors.py
